DB rollback, Add new migration, validate price to decimal, Add RecipeSpec

+++ db/migrate/20220715070931_add_price_to_recipes.rb
@@ -1,5 +1,5 @@
 class AddPriceToRecipes < ActiveRecord::Migration[6.1]
   def change
-    add_column :recipes, :price, :float
+    add_column :recipes, :price, :decimal
   end
 end
